ARM64: Bring up mt6983 platform and matisse device
- Switch platform defconfig to xiaomi_mt6983.config and build mt6983 dtb
- Add vendor defconfig fragments for the mt6983 platform and matisse
- Add build.config.xiaomi.matisse device build config
- Add L11 38.0A 0A DSC CMD panel driver and MT6983 btif platform header
- PM: restore legacy wakelock API via wakeup_source compat shim
- mali: re-add debugfs objects missing from the valhall build
- apusys: guard tracepoint helpers with CONFIG_TRACEPOINTS
- drm: drop legacy videolfb/fbdev code
- connectivity: switch BT/WLAN platform to 6983 and build wmt common core,
  drop stale connsyslog/EMI export code
- st21nfc: add ACPI match table
- musb: remove dropped dev_dma_ops workaround

+++ drivers/gpu/drm/mediatek/mediatek_v2/mtk_drm_crtc.c
@@ -7032,7 +7032,7 @@ static void msync_add_frame_time(struct mtk_drm_crtc *mtk_crtc,
 		fps = drm_mode_vrefresh(mode);
 		time_diff = msync_dy->record[msync_dy->record_index].time -
 			msync_dy->record[last_msync_idx].time;
-		DDPDBG("[Msync] min fps:%f, fps:%d, time_diff:%d\n", MSYNC_MIN_FPS, fps, time_diff);
+		DDPDBG("[Msync] min fps:%d, fps:%d, time_diff:%d\n", (int)MSYNC_MIN_FPS, fps, time_diff);
 		if (1000 * 1000 * 1000 / MSYNC_MIN_FPS < time_diff) {
 			msync_dy->record[msync_dy->record_index].low_frame = true;
 			DDPDBG("[Msync] low_frame = true\n");
